Add a validation webhook for authentication to providers
This webhook only validates secrets that are set with the
'createdForProviderType' label that is supposed to be set by clients,
e.g., the forlift-console-plugin.
