+++ .bazelrc
@@ -7,6 +7,7 @@ build --action_env=DEFAULT_CHANNEL=development
 
 ## Images which should be installed
 build --action_env=CONTROLLER_IMAGE=quay.io/kubev2v/forklift-controller:latest
+build --action_env=API_IMAGE=quay.io/kubev2v/forklift-api:latest
 build --action_env=MUST_GATHER_IMAGE=quay.io/kubev2v/forklift-must-gather:latest
 build --action_env=MUST_GATHER_API_IMAGE=quay.io/kubev2v/forklift-must-gather-api:latest
 build --action_env=UI_IMAGE=quay.io/kubev2v/forklift-ui:latest
