+++ operator/roles/forkliftcontroller/defaults/main.yml
@@ -88,6 +88,19 @@ ui_plugin_container_limits_memory: "800Mi"
 ui_plugin_container_requests_cpu: "100m"
 ui_plugin_container_requests_memory: "150Mi"
 
+api_image_fqin: "{{ lookup( 'env', 'API_IMAGE') or lookup( 'env', 'RELATED_IMAGE_API') }}"
+api_service_name: "{{ app_name }}-api"
+api_deployment_name: "{{ api_service_name }}"
+api_container_name: "{{ app_name }}-api"
+api_container_limits_cpu: "1000m"
+api_container_limits_memory: "1Gi"
+api_container_requests_cpu: "100m"
+api_container_requests_memory: "150Mi"
+api_tls_secret_name: "{{ api_service_name }}-serving-cert"
+api_selfsigned_issuer_name: "{{ api_service_name }}-selfsigned-issuer"
+api_issuer_name: "{{ api_service_name }}-issuer"
+api_certificate_name: "{{ api_service_name }}-selfsigned-ca"
+
 must_gather_api_image_fqin: "{{ lookup( 'env', 'MUST_GATHER_API_IMAGE') or lookup( 'env', 'RELATED_IMAGE_MUST_GATHER_API') }}"
 must_gather_api_service_name: "{{ app_name }}-must-gather-api"
 must_gather_api_deployment_name: "{{ must_gather_api_service_name }}"
