+++ operator/roles/forkliftcontroller/tasks/main.yml
@@ -80,6 +80,27 @@
       definition: "{{ lookup('template', 'controller/provider-host.yml.j2') }}"
     when: "'kubevirt.io' in api_groups"
 
+  - name: "Configure webhook certificate on K8s"
+    k8s:
+      state: present
+      definition: "{{ lookup('template', 'api/ca.yml.j2') }}"
+    when: k8s_cluster|bool
+
+  - name: "Setup api service"
+    k8s:
+      state: present
+      definition: "{{ lookup('template', 'api/service-forklift-api.yml.j2') }}"
+
+  - name: "Setup api deployment"
+    k8s:
+      state: present
+      definition: "{{ lookup('template', 'api/deployment-forklift-api.yml.j2') }}"
+
+  - name: "Setup api validating webhook configuration"
+    k8s:
+      state: present
+      definition: "{{ lookup('template', 'api/validatingwebhookconfiguration-forklift-api.yml.j2') }}"
+
   - when: feature_validation|bool
     block:
     - name: "Setup validation service"
