+++ operator/roles/forkliftcontroller/templates/api/service-forklift-api.yml.j2
@@ -1,0 +1,21 @@
+---
+apiVersion: v1
+kind: Service
+metadata:
+  annotations:
+    service.beta.openshift.io/serving-cert-secret-name: {{ api_tls_secret_name }}
+  name: {{ api_service_name }}
+  namespace: "{{ app_namespace }}"
+  labels:
+    app: {{ app_name }}
+    service: {{ api_service_name }}
+spec:
+  type: ClusterIP
+  selector:
+    app: {{ app_name }}
+    service: {{ api_service_name }}
+  ports:
+  - name: api-https
+    port: 443
+    targetPort: 8443
+    protocol: TCP
