+++ pkg/forklift-api/BUILD.bazel
@@ -5,5 +5,8 @@ go_library(
     srcs = ["api.go"],
     importpath = "github.com/konveyor/forklift-controller/pkg/forklift-api",
     visibility = ["//visibility:public"],
-    deps = ["//pkg/forklift-api/webhooks"],
+    deps = [
+        "//pkg/forklift-api/webhooks",
+        "//pkg/lib/logging",
+    ],
 )
