+++ pkg/forklift-api/webhooks/validating-webhook.go
@@ -7,10 +7,6 @@ import (
 	"github.com/konveyor/forklift-controller/pkg/forklift-api/webhooks/validating-webhook/admitters"
 )
 
-func ServeProviderCreate(resp http.ResponseWriter, req *http.Request) {
-	validating_webhooks.Serve(resp, req, &admitters.ProviderAdmitter{})
-}
-
 func ServeSecretCreate(resp http.ResponseWriter, req *http.Request) {
 	validating_webhooks.Serve(resp, req, &admitters.SecretAdmitter{})
 }
